+++ modules/StoreManagement/Database/Migrations/2021_05_20_155812_create_geopolitical_zones_table.php
@@ -18,8 +18,8 @@ class CreateGeopoliticalZonesTable extends Migration
             $table->string('name');
             $table->string('code_name');
             $table->foreignId('country_id')->constrained()->restrictOnDelete();
-            $table->float('longitude')->nullable();
-            $table->float('latitude')->nullable();
+            $table->double('longitude')->nullable();
+            $table->double('latitude')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });
