+++ modules/StoreManagement/Database/Migrations/2021_05_20_160520_create_regions_table.php
@@ -18,8 +18,8 @@ class CreateRegionsTable extends Migration
             $table->string('name');
             $table->string('code_name');
             $table->foreignId('geopolitical_zone_id')->constrained()->restrictOnDelete();
-            $table->float('longitude')->nullable();
-            $table->float('latitude')->nullable();
+            $table->double('longitude')->nullable();
+            $table->double('latitude')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });
